[feature] failpopupview 에 viewModel 데이터 할당
